Implemented support for `dump`ing to a file
Added the `-f` and `-o` options to the `dump` command.
Pass the outputFile option along to `printEnvironment()`.
If json is set or if the output file option is set, it will
emit json instead of console-formatted data.

Added node 4.* and 5.* to the travis test targets, and switched
to using Travis's newer infra.